requirements.txt
flask
flask-sqlalchemy==2.0
flask-wtf
flask-babel
flask-mail
flask-script
flask-compress
flask-oauthlib
flask-socketio
flask-bcrypt
flask-redis
flask-marshmallow
sqlalchemy==1.0.8
psycopg2
pillow
oauth
redis
python-slugify
marshmallow
marshmallow-sqlalchemy
werkzeug==0.10.4

